add: jwtUtils

+++ server/api/todoList/todo/controller.ts
@@ -2,8 +2,13 @@ import { todoUseCase } from 'domain/todolist/useCase/todoUseCase';
 import { defineController } from './$relay';
 
 export default defineController(() => ({
-  get: async () => {
+  get: async ({ headers }) => {
     try {
+      const todoUserId = parseInt(headers['x-todo-user-id'] as string, 10);
+      if (isNaN(todoUserId)) {
+        return { status: 400, body: { error: 'Invalid user ID' } };
+      }
+
       const todos = await todoUseCase.getTodo();
       return {
         status: 200,
@@ -17,9 +22,13 @@ export default defineController(() => ({
       };
     }
   },
-  post: async ({ body }) => {
+  post: async ({ body, headers }) => {
     console.log('post', body);
     try {
+      const todoUserId = parseInt(headers['x-todo-user-id'] as string, 10);
+      if (isNaN(todoUserId)) {
+        return { status: 400, body: { error: 'Invalid user ID' } };
+      }
       const todo = await todoUseCase.createTodo(body.title, body.notes ?? '');
       return {
         status: 201,
@@ -33,9 +42,13 @@ export default defineController(() => ({
       };
     }
   },
-  put: async ({ body }) => {
+  put: async ({ body, headers }) => {
     console.log('put', body);
     try {
+      const todoUserId = parseInt(headers['x-todo-user-id'] as string, 10);
+      if (isNaN(todoUserId)) {
+        return { status: 400, body: { error: 'Invalid user ID' } };
+      }
       const todo = await todoUseCase.updateTodo(body.id, body.title, body.notes ?? '');
       if (todo) {
         return { status: 200, body: todo };
@@ -50,8 +63,12 @@ export default defineController(() => ({
       };
     }
   },
-  delete: async ({ body }) => {
+  delete: async ({ body, headers }) => {
     try {
+      const todoUserId = parseInt(headers['x-todo-user-id'] as string, 10);
+      if (isNaN(todoUserId)) {
+        return { status: 400, body: { error: 'Invalid user ID' } };
+      }
       const result = await todoUseCase.deleteTodo(body.id);
       if (result) {
         return { status: 204, body: undefined };
